Refactor into ES modules and add bundler
