+++ assets/js/main.js
@@ -1,0 +1,4 @@
+import { init } from './tasks.js';
+import './filters.js';
+
+document.addEventListener('DOMContentLoaded', init);
