+++ vite.config.js
@@ -1,0 +1,14 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  build: {
+    outDir: 'dist',
+    rollupOptions: {
+      input: 'assets/js/main.js',
+      output: {
+        entryFileNames: 'assets/js/bundle.js',
+        assetFileNames: 'assets/[name][extname]'
+      }
+    }
+  }
+});
